requirements.txt
playwright>=1.47
Pillow>=10.3
typer>=0.12
